Homework_3: screenshots, game functionality and comments w/ labeled OOP concepts

+++ src/main/java/CharacterInterface.java
@@ -1,8 +1,9 @@
+import java.util.List;
+
 public interface CharacterInterface {
 
-    public int getHealth();
-    public void subtractHealth(int val);
+    public double getHealth();
+    public void subtractHealth(double val);
     public String getName();
-
     public boolean isAlive();
 }
